Refactor klasy Listenera dla figury Joe 1, dodanie kolejnego dla Joe 2
