+++ src/main/java/edu/iis/powp/events/predefine/figureslisteners/SelectTestFigureOptionJoe2Listener.java
@@ -1,0 +1,19 @@
+package edu.iis.powp.events.predefine.figureslisteners;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+import edu.iis.client.plottermagic.preset.FiguresJoe;
+import edu.iis.powp.app.Application;
+import edu.iis.powp.app.DriverManager;
+
+public class SelectTestFigureOptionJoe2Listener implements ActionListener
+{
+
+    @Override
+    public void actionPerformed(ActionEvent e)
+    {
+
+        FiguresJoe.figureScript2(Application.getComponent(DriverManager.class).getCurrentPlotter());
+    }
+}
