+++ src/main/java/edu/iis/powp/events/predefine/figureslisteners/SelectTestFigureOptionJoeListener.java
@@ -1,4 +1,4 @@
-package edu.iis.powp.events.predefine;
+package edu.iis.powp.events.predefine.figureslisteners;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -7,12 +7,13 @@ import edu.iis.client.plottermagic.preset.FiguresJoe;
 import edu.iis.powp.app.Application;
 import edu.iis.powp.app.DriverManager;
 
-public class SelectTestFigureOptionListener implements ActionListener
+public class SelectTestFigureOptionJoeListener implements ActionListener
 {
 
     @Override
     public void actionPerformed(ActionEvent e)
     {
         FiguresJoe.figureScript1(Application.getComponent(DriverManager.class).getCurrentPlotter());
+        
     }
 }
