+++ src/test/java/edu/iis/powp/gui/TestPlotSoftPatterns.java
@@ -13,7 +13,8 @@ import edu.iis.powp.app.Context;
 import edu.iis.powp.app.DriverManager;
 import edu.iis.powp.appext.ApplicationWithDrawer;
 import edu.iis.powp.events.predefine.SelectChangeVisibleOptionListener;
-import edu.iis.powp.events.predefine.SelectTestFigureOptionListener;
+import edu.iis.powp.events.predefine.figureslisteners.SelectTestFigureOptionJoe2Listener;
+import edu.iis.powp.events.predefine.figureslisteners.SelectTestFigureOptionJoeListener;
 import edu.kis.powp.drawer.panel.DefaultDrawerFrame;
 import edu.kis.powp.drawer.panel.DrawPanelController;
 
@@ -28,9 +29,10 @@ public class TestPlotSoftPatterns
 	 * @param context Application context.
 	 */
 	private static void setupPresetTests(Context context) {
-	    SelectTestFigureOptionListener selectTestFigureOptionListener = new SelectTestFigureOptionListener();
-		
-		context.addTest("Figure Joe 1", selectTestFigureOptionListener);	        
+	    SelectTestFigureOptionJoeListener selectTestFigureOptionJoeListener = new SelectTestFigureOptionJoeListener();
+	    SelectTestFigureOptionJoe2Listener selectTestFigureOptionJoe2Listener = new SelectTestFigureOptionJoe2Listener();
+		context.addTest("Figure Joe 1", selectTestFigureOptionJoeListener);
+		context.addTest("Figure Joe 2", selectTestFigureOptionJoe2Listener);
 	}
 
 	/**
